api/requirements.txt
fastapi[standard]==0.115.6
passlib==1.7.4
pydantic==2.10.5
PyJWT==2.10.1
python-dotenv==1.0.1
Requests==2.32.3
sqlmodel==0.0.22
argon2_cffi==23.1.0
